Add copy member info on hotkey
